+++ src/routes/characters/+page.svelte
@@ -3,7 +3,7 @@
 	import type { CharactersResponse } from '$lib/pocketbase-types';
 	import { toastShow } from '$lib/toast';
 	import { cloneCharacter, deleteCharacter, getAllCharacters } from '$models/character';
-	import { ProgressRadial, type PopupSettings, popup, type ModalSettings, getModalStore } from '@skeletonlabs/skeleton';
+	import { ProgressRadial, type PopupSettings, popup, type ModalSettings, getModalStore, getToastStore } from '@skeletonlabs/skeleton';
 	import { flip } from 'svelte/animate';
 	import { crossfade, fade } from 'svelte/transition';
 	import { quintOut } from 'svelte/easing';
@@ -12,6 +12,7 @@
 	import CharacterItem from '$lib/components/characters/CharacterItem.svelte';
 
 	const modalStore = getModalStore()
+	const toastStore = getToastStore()
 
 	const charOperationsMenu: PopupSettings = {
 		event: 'focus-click',
@@ -34,7 +35,7 @@
 	async function cloneSelectedCharacter() {
 		const character = await cloneCharacter(operationsOnCharacterId)
 
-		toastShow(`Character <span class="text-secondary-100">${character.name}</span> has been cloned`)
+		toastShow(`Character <span class="text-secondary-100">${character.name}</span> has been cloned`, toastStore)
 
 		myCharacters = [...myCharacters, character]
 	}
@@ -53,7 +54,7 @@
 
 						await deleteCharacter(character.id)
 
-						toastShow(`Character <span class="text-error-900">${character.name}</span> has been removed`)
+						toastShow(`Character <span class="text-error-900">${character.name}</span> has been removed`, toastStore)
 
 						// Update reactively so the element refreshes.
 						myCharacters = myCharacters.filter(c => c.id!== character.id)